refactor: migrate repositories and services to new prisma client setup
