+++ apps/api/src/notice/notice.controller.ts
@@ -27,7 +27,7 @@ import {
 import { NoticeService } from './notice.service';
 import { GeneralNoticeListDto } from './dto/res/generalNotice.dto';
 import { GetAllNoticeQueryDto } from './dto/req/getAllNotice.dto';
-import { User } from '@prisma/client';
+import { User } from '@generated/prisma/client';
 import { ExpandedGeneralNoticeDto } from './dto/res/expandedGeneralNotice.dto';
 import { CreateNoticeDto } from './dto/req/createNotice.dto';
 import { ForeignContentDto } from './dto/req/foreignContent.dto';
