+++ apps/api/src/notice/types/noticeFullContent.ts
@@ -1,4 +1,4 @@
-import { Prisma } from '@prisma/client';
+import { Prisma } from '@generated/prisma/client';
 
 export type NoticeFullContent = Prisma.NoticeGetPayload<{
   include: {
