+++ apps/api/src/user/decorator/get-user.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { User } from '@prisma/client';
+import { User } from '@generated/prisma/client';
 
 export const GetUser = createParamDecorator(
   (_data, ctx: ExecutionContext): User | undefined => {
