+++ apps/api/src/user/guard/idp.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-http-bearer';
 import { UserService } from '../user.service';
-import { User } from '@prisma/client';
+import { User } from '@generated/prisma/client';
 import { InfoteamIdpService } from '@lib/infoteam-idp';
 import { UserInfo } from '@lib/infoteam-idp/types/userInfo.type';
