+++ apps/crawler/src/user/user.repository.ts
@@ -1,7 +1,7 @@
 import { Loggable } from '@lib/logger/decorator/loggable';
 import { PrismaService } from '@lib/prisma';
 import { Injectable } from '@nestjs/common';
-import { User } from '@prisma/client';
+import { User } from '@generated/prisma/client';
 import { v4 as uuid } from 'uuid';
 
 @Loggable()
