+++ apps/crawler/src/user/user.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { UserRepository } from './user.repository';
-import { User } from '@prisma/client';
+import { User } from '@generated/prisma/client';
 import { Loggable } from '@lib/logger/decorator/loggable';
 
 @Loggable()
